+++ libTracer/Scene/Computations.cs
@@ -6,7 +6,7 @@ namespace libTracer.Scene
 {
     public class Computations
     {
-        public Single Time { get; set; }
+        public Double Time { get; set; }
         public Shape Object { get; set; }
         public TPoint Point { get; set; }
         public TPoint OverPoint { get; set; }
@@ -15,7 +15,7 @@ namespace libTracer.Scene
         public TVector NormalV { get; set; }
         public TVector ReflectV { get; set; }
         public Boolean Inside { get; set; }
-        public Single N1 { get; set; }
-        public Single N2 { get; set; }
+        public Double N1 { get; set; }
+        public Double N2 { get; set; }
     }
 }
